# Build only Debug and Release configurations in TestBuildFBuild