This change speeds up `BuildFBuild` tests by minimizing amount of build configurations while maintaining almost the same amount of code exposed to compilers. Resulting speedup is especially significant on Linux (from 459 seconds down to 109 seconds in my case).

1. Sanitizer configurations are excluded because they have the same defines as Release configuration. Also just building code with sanitizers doesn't uncover any bugs by itself.
2. Profile configuration has a different set of defines than Debug and Release configurations but almost all code is already covered by them. (There are only trivial code behind `defined( RELEASE ) &&
   defined( PROFILING_ENABLED )` condition.)