@ffulin 2018-05-26
This change makes sense - thanks!

I realize that if you are pulling the repo from git, then because all the files are writeable all the files will be isolated from Unity, so you'd be getting *really* slow compile times:

| | Isolated | Unity |
| Before | 812s | 93s |
| With This Pull Request | 564s | 68s |

I've made a change to the bff (https://github.com/fastbuild/fastbuild/commit/8c792e558730c55e04250f47dfe45214c0210882) so that when FASTBuild is compiling itself, it will be done entirely in Unity, so now we get the faster time (68s above) in all cases